rename `parser` to avoid name collision w/ stdlib
