+++ guba_parser.py
